Implementacion de PSR4 | todas las consultas ahora se hacen desde class Query

+++ backend/v1/users/history/HistoryController.php
@@ -1,28 +1,27 @@
 <?php
 
+namespace V1\Users\History;
+
+use Db\Querys;
+use Exception;
+use Tools\JsonResponse;
+
 class HistoryController extends Querys
 {
    private const SET = 'history_id, user_id, video_id, history_views, update_date';
+   protected const TIME = 'Y-m-d H:i:00';
 
    public static function index()
    {
       $historyQuery = new Querys('history');
-      $videoQuery = new Querys('videos');
 
       $arrayHistory = $historyQuery->select('user_id', $_GET['id'], 'video_id, history_views, update_date');
-      if ($arrayHistory) {
-         foreach ($arrayHistory as $history) {
-            $video_id = $history->video_id;
-            $arrayVideo = $videoQuery->select('video_id', $video_id, 'name, link, provider_logo');
+      if ($arrayHistory == false) throw new Exception('not found resourse', 404);
 
-            $_history = $arrayVideo[0];
-            $_history->views = $history->history_views;
-            $_history->date = $history->update_date;
-
-            $_arrayHistory[] = $_history;
-         }
-      } else throw new Exception('not found resourse', 404);
-
+      foreach ($arrayHistory as $history) {
+         $history_data = self::getHistoryData($history);
+         $_arrayHistory[] = $history_data;
+      }
       JsonResponse::read('history', $_arrayHistory);
    }
 
@@ -33,18 +32,9 @@ class HistoryController extends Querys
 
       $history_id = $_GET['id'] . $_GET['id_2'];
       $arrayHistory = $historyQuery->select('history_id', $history_id, 'video_id, history_views, update_date');
+      if ($arrayHistory == false) throw new Exception('not found resourse', 404);
 
-      if ($arrayHistory) {
-         $history = $arrayHistory[0];
-         $video_id = $history->video_id;
-         $arrayVideo = $videoQuery->select('video_id', $video_id, 'name, link, provider_logo');
-
-         $_history = $arrayVideo[0];
-         $_history->views = $history->history_views;
-         $_history->date = $history->update_date;
-
-      } else throw new Exception('not found resourse', 404);
-
+      $_history = self::getHistoryData($arrayHistory[0]);
       JsonResponse::read('history', $_history);
    }
 
@@ -55,21 +45,27 @@ class HistoryController extends Querys
       $history_id = $_GET['id'] . $_GET['id_2'];
       $arrayHistory = $historyQuery->select('history_id', $history_id, 'history_views');
 
+      date_default_timezone_set('America/Caracas');
       if ($arrayHistory) {
          $history = $arrayHistory[0];
-         $_arrayHistory['history_views'] = ++$history->history_views;
+
+         $_arrayHistory = [
+            'history_views' => ++$history->history_views,
+            'update_date' => date(self::TIME)
+         ];
          $historyQuery->update('history_id', $history_id, $_arrayHistory);
+
       } else {
-         $_history = [
+         $_arrayHistory = [
             'history_id' => $_GET['id'] . $_GET['id_2'],
             'user_id' => $_GET['id'],
-            'video_id' => $_GET['id_2']
+            'video_id' => $_GET['id_2'],
+            'history_views' => 1,
+            'update_date' => date(self::TIME)
          ];
-         $historyQuery->insert($_history);
+         $historyQuery->insert($_arrayHistory);
       }
-
-      $history = $historyQuery->select('history_id', $history_id, 'history_id, history_views, create_date, update_date');
-      JsonResponse::created('history', $history);
+      JsonResponse::created('history', $_arrayHistory);
    }
 
    public static function destroy()
@@ -77,7 +73,24 @@ class HistoryController extends Querys
       $historyQuery = new Querys('history');
 
       $history_id = $_GET['id'] . $_GET['id_2'];
+      $arrayHistory = $historyQuery->select('history_id', $history_id);
+      if ($arrayHistory == false) throw new Exception('not found resourse', 404);
+
       $historyQuery->delete('history_id', $history_id);
       JsonResponse::removed();
+   }
+
+   protected static function getHistoryData($history)
+   {
+      $videoQuery = new Querys('videos');
+
+      $video_id = $history->video_id;
+      $arrayVideo = $videoQuery->select('video_id', $video_id, 'video_id, name, link, provider_logo');
+
+      $_history = $arrayVideo[0];
+      $_history->views = $history->history_views;
+      $_history->date = $history->update_date;
+
+      return $_history;
    }
 }
